Version 0.9.2
* Fix issue with foreign path separator
* Move files to src/ and update import
* Add Flatpak manifest
* Add helper script for Flatpak build
* Format files
* Add arch name to bundle
* Add creation of $BUILD_DIR
* Add notes
* Resolve number of script's arguments
* Add mnemonic options
* Add test for Docker binary
* Add hints for specifing a command with a word
* Update promt message
* Split instructions into 2 docs

+++ scripts/build-lnx.sh
@@ -10,43 +10,43 @@ echo "Project home: $PWD";
 # Get a platform architecture
 ARCH=$(uname -m);
 BIN_NAME="columbo-$ARCH";
-echo "Detected platform architecture: $ARCH";
+echo "Platform architecture: $ARCH";
 
-# Check the specified mode
+# Resolve build mode
 if [ $# -gt 0 ];
 then
 	MODE=$1;
 else
-  if ! [ -v MODE ];   # If MODE is not defined in the environment (is used in Docker image)
-  then
-    echo "
-Build mode was not specified as an argument or as an environment variable MODE. Choose build mode:
-  [1] onedir (one executable file and a folder with supplementary files, works faster)
+	if ! [ -v MODE ];   # If MODE is not defined in the environment (used in Docker image)
+	then
+		echo "Build mode is not specified. Choose build mode:
+  [1] onedir  (one executable file and a folder with supplementary files, works faster)
   [2] onefile (all in one executable file, works slower)
+(Build mode can be provided as an script argument or as an environment variable MODE)
 ";
-
-    read -p "Enter a number of an option or just press ENTER for default 'onedir' option: " OPTION;
-    case $OPTION in
-      1 )
-        MODE=onedir;
-      ;;
-      2 )
-        MODE=onefile;
-      ;;
-      * )
-        MODE=onedir;
-      ;;
-    esac
-  fi
+		read -p "Enter [1-2] or just press ENTER for the default 'onedir' option: " OPTION;
+		case $OPTION in
+			1 )
+				MODE=onedir;
+			;;
+			2 )
+				MODE=onefile;
+			;;
+			* )
+				MODE=onedir;
+			;;
+		esac
+	fi
 fi
-echo "Mode: $MODE";
+echo "Build mode: $MODE";
 
+# Resolve build tag
 if [ -v TAG ];
 then
 	echo "Build tag: $TAG";
 	BIN_NAME="$BIN_NAME-$TAG";
 else
-	echo "Build tag is not defined";
+	echo "Build tag is not specified";
 fi
 
 # Create executable file
@@ -58,23 +58,23 @@ python3 -m PyInstaller \
 	--windowed \
 	--noconfirm \
 	--clean \
-	--add-data="./gui:./gui" \
-	--add-data="./icons:./icons" \
-	--icon=icons/search.ico \
-	main.py;
+	--add-data="./src/gui:./src/gui" \
+	--add-data="./resources/icons:./resources/icons" \
+	--icon=resources/icons/search.ico \
+	src/main.py;
 
 # Delete temp folder
 echo "+ Remove 'build' folder";
 rm -R build;
-echo "+ Remove .spec";
+echo "+ Remove '.spec' file";
 rm $BIN_NAME.spec;
 
 # Create an archive with binary
-if [ "$MODE" = "onedir" ];
+if [ "$MODE" = "onedir" ] && [ -d "dist/$BIN_NAME" ];
 then
   echo "+ Pack a folder with binary...";
   cd dist || exit 1;
-  tar -czf "$BIN_NAME".tar.gz "$BIN_NAME"
+  tar -czf "$BIN_NAME".tar.gz "$BIN_NAME";
   rm -R "$BIN_NAME";
   echo "+ Columbo is packed to dist/$BIN_NAME.tar.gz";
 fi
